trainer/requirements.txt
Flask==3.0.2
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.1
numpy==1.26.4
pillow==10.2.0
requests==2.31.0